Update 'make-img-linux.sh' to be a bit more error tolerant

+++ tests/make-img-linux.sh
@@ -1,8 +1,17 @@
-#!/bin/bash
-# please install dosfstools
+#!/usr/bin/env bash
 
+set -eax
 
-set -ex
+if ! command -v "mkfs.fat" &>/dev/null; then
+    echo "Please install 'dosfstools'"
+    exit 1
+fi
+
+BRANCH_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && cd .. && pwd)
+TARGET_DIR="${BRANCH_DIR}/dist"
+mkdir -p "${TARGET_DIR}"
+cd "${TARGET_DIR}" || exit 2
+
 # 1Mb
 dd if=/dev/zero of=fat12.img bs=1048576 count=1
 # 1MB
@@ -13,4 +22,3 @@ dd if=/dev/zero of=fat32.img bs=1048576 count=64
 mkfs.fat -F 12 fat12.img
 mkfs.fat -F 16 fat16.img
 mkfs.fat -F 32 fat32.img
-
